Supress errors about file not found
When we're waiting for browserstack to start, the log file that is
checked might not exist yet resulting in a "file not found error".

Adding '-s' to grep supress those errors.

One consern though is that if the file is missing because of other
problems, this might be hard to debug.

Maybe there stderr should be redirected to a file so it can be debugged.

Something like:
```
while ! grep -s "${DONE_MSG}" ${BROWSER_STACK_LOG_FILE} 2>
grep_errors.log > /dev/null; do
```