Cache the selenium-node image in cluster
Sets the `imagePullPolicy` to `IfNotPresent` instead of `Always`. 
This should allow caching in the cluster so that k8s does not have to make a request to the docker hub every time a node is to be created, thereby lowering overhead and network IO.
Apart from this, I have observed that scaling up nodes might present issues for docker hub - likely resource constraints from their side.

To avoid readiness probe errors from the nodes, it could be beneficial setting up the timeout from 1 second to 5. There isn't a lot of room for error with just one second; at the least, it's been observed that this made it slightly more stable from our side.