Cache the selenium-node image in cluster (#764)
* Cache the selenium-node image in cluster

Sets the `imagePullPolicy` to `IfNotPresent` instead of `Always`. 
This should allow caching in the cluster so that k8s does not have to make a request to the docker hub every time a node is to be created, thereby lowering overhead and network IO.
Apart from this, I have observed that scaling up nodes might present issues for docker hub - likely resource constraints from their side.

To avoid readiness probe errors from the nodes, it could be beneficial setting up the timeout from 1 second to 5. There isn't a lot of room for error with just one second; at the least, it's been observed that this made it slightly more stable from our side.

* Add ImagePullPolicy as an enum

* Add ImagePullPolicy to the config. Still needs to be read as a parameter if provided, or default to "Always". Right now, it is hard-coded as "Always".

* Fix timeout settings(mistakenly put 10 instead of 5)

* Fix auto-formatting of tabs

* Fix white-space auto-formatting

* Default imagePullPolicy to Always if environment "ZALENIUM_KUBERNETES_IMAGE_PULL_POLICY" is not set.

* Whitespace in ENV variable.

* Fix IntelliJ auto formatting(whitespace issues)

* Simplify logic for setting ENV variable for "ZALENIUM_KUBERNETES_IMAGE_PULL_POLICY"