Start stop video by cookie (#1138)
* Add feature to start or stop video recording using cookie

Per #490, video recording can be started, if disabled, or stopped, if enabled, by sending cookie named 'zaleniumVideo'. This is beneficial when multiple tests run in the same session.

Additionally to assist with identifying tests in the dashboard, a cookie named 'zaleniumTestName' can be passed in to set the test name. This name will appear in the dashboard. If name already exists, a count will be appended to the name.

* Add feature to start or stop video recording using cookie Part 2

Adding documentation and integration test along with small change in duplicate file naming logic. Also, revert back the logic of the keepVideoAndLogs method

* Add feature to start or stop video recording using cookie Part 3

Add wait for items to appear in the dashboard for the IT

Co-authored-by: Chanatan Charnkijtawarush <ccharnkij@gmail.com>