Add feature to start or stop video recording using cookie (#1085)
* Add feature to start or stop video recording using cookie

Per #490, video recording can be started, if disabled, or stopped, if enabled, by sending cookie named 'zaleniumVideo'. This is beneficial when multiple tests run in the same session.

Additionally to assist with identifying tests in the dashboard, a cookie named 'zaleniumTestName' can be passed in to set the test name. This name will appear in the dashboard. If name already exists, a count will be appended to the name.

* Making the test more complete

* Add feature to start or stop video recording using cookie

Per #490, video recording can be started, if disabled, or stopped, if enabled, by sending cookie named 'zaleniumVideo'. This is beneficial when multiple tests run in the same session.

Additionally to assist with identifying tests in the dashboard, a cookie named 'zaleniumTestName' can be passed in to set the test name. This name will appear in the dashboard. If name already exists, a count will be appended to the name.

* Adding TestNG groups

Co-authored-by: Chanatan Charnkijtawarush <C.Charnkijtawarush@allegiantair.com>
Co-authored-by: Diego Molina <diemol@gmail.com>